+++ packages/@coorpacademy-components/src/organism/get-the-app/index.js
