+++ packages/@coorpacademy-components/src/organism/get-the-app/test/fixtures/default.js
